Eventually implemented correct id(.).
This should assure the correctness e.g. of pickle and copy.deepcopy.
[notice that cPickle was unaffected]

In the long run at least copy should use directly an identity mapping instead
of id() [this should be discussed with Python-Dev together with the idea
of deprecating/stear away from  id() in favor of offering (weak) identity mappings to
users when a mapping is what is needed].

So this change will be less important, still as long as Python has a id(.)
semantics is better implemented correctly.

ids are still useful for tracebility/debugging.

Converted string reprs (toString results) to call Py.idstr(.) instead
of directly "at "+Py.id(.) which was implemented simply as
System.identityHashCode.

The final form of the results of Py.idstr(.) is undecided.

On a less serious note:
>>> class A: pass
...
>>> id(A())
1
>>> A()
<__main__.A instance 2>
>>>

yes the new ids look funny.

+++ org/python/core/__builtin__.java
@@ -404,7 +404,7 @@ public class __builtin__ implements ClassDictInit
         return o.__hex__();
     }
 
-    public static int id(PyObject o) {
+    public static long id(PyObject o) {
         return Py.id(o);
     }
